update: modal database

+++ src/controllers/ProductController.ts
@@ -9,6 +9,7 @@ const getAllProducts = async (
 ) => {
   try {
     const products = await productService.getAllProducts();
+
     return response.ok(res, products, "Get all products successfully");
   } catch (error) {
     next(error);
@@ -21,7 +22,33 @@ const createProduct = async (
   next: NextFunction,
 ) => {
   try {
-    const product = await productService.createProduct(req.body);
+    const files = req.files as Express.Multer.File[];
+
+    let uploadedImages: string[] = [];
+
+    if (files?.length) {
+      uploadedImages = await Promise.all(files.map((file) => file.path));
+    }
+
+    const variants = JSON.parse(req.body.variants);
+
+    // gán image theo variant index
+    variants.forEach((variant: any, index: number) => {
+      variant.images = uploadedImages[index]
+        ? [
+            {
+              imageUrl: uploadedImages[index],
+              isPrimary: true,
+            },
+          ]
+        : [];
+    });
+
+    const product = await productService.createProduct({
+      ...req.body,
+      variants,
+    });
+
     return response.ok(res, product, "Product created successfully");
   } catch (error) {
     next(error);
@@ -34,11 +61,18 @@ const getProductById = async (
   next: NextFunction,
 ) => {
   try {
-    const id = parseInt(req.params.id as string);
+    const id = Number(req.params.id);
+
+    if (!id) {
+      return response.badRequest(res, null, "Invalid product id");
+    }
+
     const product = await productService.getProductById(id);
+
     if (!product) {
       return response.notFound(res, null, "Product not found");
     }
+
     return response.ok(res, product, "Product found successfully");
   } catch (error) {
     next(error);
@@ -51,8 +85,14 @@ const updateProduct = async (
   next: NextFunction,
 ) => {
   try {
-    const id = parseInt(req.params.id as string);
+    const id = Number(req.params.id);
+
+    if (!id) {
+      return response.badRequest(res, null, "Invalid product id");
+    }
+
     const product = await productService.updateProduct(id, req.body);
+
     return response.ok(res, product, "Product updated successfully");
   } catch (error) {
     next(error);
@@ -65,12 +105,20 @@ const deleteProduct = async (
   next: NextFunction,
 ) => {
   try {
-    const id = parseInt(req.params.id as string);
+    const id = Number(req.params.id);
+
+    if (!id) {
+      return response.badRequest(res, null, "Invalid product id");
+    }
+
     const product = await productService.getProductById(id);
+
     if (!product) {
       return response.notFound(res, null, "Product not found");
     }
+
     await productService.deleteProduct(id);
+
     return response.ok(res, null, "Product deleted successfully");
   } catch (error) {
     next(error);
@@ -83,8 +131,14 @@ const getProductsByCategoryId = async (
   next: NextFunction,
 ) => {
   try {
-    const categoryId = parseInt(req.params.categoryId as string);
+    const categoryId = Number(req.params.categoryId);
+
+    if (!categoryId) {
+      return response.badRequest(res, null, "Invalid category id");
+    }
+
     const result = await productService.getProductsByCategoryId(categoryId);
+
     return response.ok(res, result, "Get products by category id successfully");
   } catch (error) {
     next(error);
